requirement.txt
# Jupyter & Data Science
jupyter
ipykernel
pandas
numpy
plotly

matplotlib

# LangChain Core (moved to pyproject.toml)
# langchain
# langgraph
# streamlit
# python-dotenv
# chromadb

# LangChain Extensions
langchain_community
langchainhub
langchain-text-splitters
faiss-cpu

# LLM Providers
langchain-openai
langchain-anthropic
langchain-groq
langchain-google-genai
langchain-ollama
google-cloud-aiplatform

# Database Drivers (moved to pyproject.toml)
# pyodbc
# pymysql
# psycopg2-binary


